refactor OrgDetails into a feature specific component, use 'Org' as feature and leave 'OrgDetails' as a dumb component in charge of rendering
