+++ client/app/features/org/Org.spec.tsx
@@ -1,30 +1,23 @@
-import { render, screen } from '@testing-library/react';
 import { setupServer } from 'msw/node';
 import { afterAll, afterEach, beforeAll, describe, expect, it } from 'vitest';
-import { cleanup } from '~/mocks';
+import { cleanup, renderWithProviders } from '~/mocks';
 import { mockSiteEndpoints } from '~/mocks/handlers';
-import { OrgDetails, orgDetailsLoader } from './OrgDetails';
+import { Org, orgLoader } from './Org';
 
 const server = setupServer(...mockSiteEndpoints);
 afterEach(() => cleanup());
 beforeAll(() => server.listen());
 afterAll(() => server.close());
 
-describe('OrgDetails Component', () => {
-  it('renders the organization details heading', () => {
-    render(<OrgDetails />);
-    const headingElement = screen.getByText(/Organization Details/i);
-    expect(headingElement).toBeInTheDocument();
-  });
-
+describe('Org feature', () => {
   it('renders without crashing', () => {
-    const { container } = render(<OrgDetails />);
+    const { container } = renderWithProviders(<Org />);
     expect(container).toBeDefined();
   });
-  describe('orgDetailsLoader', () => {
+  describe('org loader function', () => {
     it('returns null when org query parameter is missing', async () => {
       const request = new Request('http://localhost/org');
-      const result = await orgDetailsLoader({ request, params: {} });
+      const result = await orgLoader({ request, params: {} });
       expect(result).toBeNull();
     });
   });
